requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
Pillow>=11.2.0
